更新zip规则

+++ 7z_all_files.sh
@@ -1,25 +1,31 @@
 #!/system/bin/sh
 
 current_dir=$(pwd)
-temp_dirs=("/tmp" "/var/tmp" "/Temp" "/Users/*/Library/Caches" "/storage/emulated/0/Android/data/bin.mt.plus/temp")
+zip7z="/data/local/tmp/7zzs"
+temp_dirs=("/tmp" "/temp" "/Temp" "/TEMP" "/TMP" "/Android/data")
 for dir in "${temp_dirs[@]}"; do
-    if [[ $current_dir == $dir* ]]; then
+    if [[ $current_dir == *"$dir"* ]]; then
         echo "当前目录是临时目录或其子目录。请解压到其他目录再执行脚本"
         echo "当前目录是临时目录或其子目录。请解压到其他目录再执行脚本"
-        echo "当前目录是临时目录或其子目录。请解压到其他目录再执行脚本，并且使用root权限执行脚本"
+        echo "当前目录是临时目录或其子目录。请解压到其他目录再执行脚本"
         exit 0
     fi
 done
+if [ "$(whoami)" != "root" ]; then
+   echo "此脚本必须以root权限运行。请使用root用户身份运行此脚本。" >&2
+   exit 1
+fi
+echo "脚本正在以root权限运行。"
 cp "$current_dir/prebuilts/7zzs" "/data/local/tmp/"
-chmod 777 "/data/local/tmp/7zzs"
-/data/local/tmp/7zzs a -r -mx=9 "$current_dir"/output.7z "$current_dir"/files/* 
+chmod 777 "$zip7z"
+$zip7z a -r -mx=9 "$current_dir"/output.7z "$current_dir"/files/*
 if [ $? -eq 0 ]; then
     echo "Successfully created archive: $current_dir/output.7z"
 else
     echo "Failed to create archive for directory: output.7z"
 fi
 rm -rf "$current_dir"/files/*
-/data/local/tmp/7zzs a -r "$current_dir"/ARMIAS.zip "$current_dir"/*
+$zip7z a -r "$current_dir"/ARMIAS.zip "$current_dir"/*
 if [ $? -eq 0 ]; then
     echo "Successfully created archive: $current_dir/ARMIAS.zip"
 else
@@ -36,5 +42,8 @@ rm -f "$current_dir"/service.sh
 rm -f "$current_dir"/output.7z
 rm -f "$current_dir"/backup_all_modules.sh
 rm -f "$current_dir"/module.prop
-rm -f "/data/local/tmp/7zzs"
-(sleep 2; rm -f "$0") &
+rm -f "$zip7z"
+(
+    sleep 2
+    rm -f "$0"
+) &
